+++ src/main/scala/com/ironcorelabs/IronSdk.scala
@@ -106,6 +106,17 @@ trait IronSdk[F[_]] {
   def userCreate(jwt: String, password: String, options: UserCreateOpts): F[UserCreateResult]
 
   /**
+   * Rotate the current user's private key, but leave the public key the same.
+   * There's no black magic here! This is accomplished via multi-party computation with the
+   * IronCore webservice.
+   *
+   * @param password password to unlock the current user's master private key
+   *
+   * @return The (encrypted) updated private key and associated metadata
+   */
+  def userRotatePrivateKey(password: String): F[UserUpdatePrivateKeyResult]
+
+  /**
    * Accesses advanced SDK operations.
    *
    * @return an instance of the [[IronSdkAdvanced]]
@@ -116,6 +127,31 @@ trait IronSdk[F[_]] {
 object IronSdk {
   import cats.effect.Sync
   import cats.implicits._
+
+  /**
+   * Initialize IronSdk with a device. Verifies that the provided user/segment exists and the provided device
+   * keys are valid and exist for the provided account.
+   *
+   * @param deviceContext device context used to initialize the IronSdk with a set of device keys
+   * @return an instance of the IronSdk
+   */
+  def initialize[F[_]](deviceContext: DeviceContext)(implicit syncF: Sync[F]): F[IronSdk[F]] =
+    deviceContext.toJava.map(com.ironcorelabs.sdk.IronSdk.initialize).map(IronSdkSync(_))
+
+  /**
+   * Initialize IronSdk with a device. Verifies that the provided user/segment exists and the provided device
+   * keys are valid and exist for the provided account.
+   * After initialization, checks whether the calling user's private key needs rotation and rotates it
+   * if necessary.
+   *
+   * @param deviceContext device context used to initialize the IronSdk with a set of device keys
+   * @param password password used to encrypt and escrow the user's private master key
+   * @return an instance of the IronSdk
+   */
+  def initializeAndRotate[F[_]](deviceContext: DeviceContext, password: String)(
+    implicit syncF: Sync[F]
+  ): F[IronSdk[F]] =
+    deviceContext.toJava.map(com.ironcorelabs.sdk.IronSdk.initializeAndRotate(_, password)).map(IronSdkSync(_))
 
   /**
    * Generate a new device for the user specified in the signed JWT.