+++ src/main/scala/com/ironcorelabs/IronSdkAdvancedFuture.scala
@@ -4,8 +4,7 @@ import scala.concurrent.Future
 import scodec.bits.ByteVector
 import cats.effect.IO
 
-case class IronSdkAdvancedFuture(deviceContext: DeviceContext) extends IronSdkAdvanced[Future] {
-  val underlying = IronSdkAdvancedSync[IO](deviceContext)
+case class IronSdkAdvancedFuture(underlying: IronSdkAdvanced[IO]) extends IronSdkAdvanced[Future] {
 
   def documentEncryptUnmanaged(data: ByteVector, options: DocumentEncryptOpts): Future[DocumentEncryptUnmanagedResult] =
     underlying.documentEncryptUnmanaged(data, options).unsafeToFuture